backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.0
apscheduler==3.10.4
psycopg2-binary==2.9.10
asyncpg==0.30.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
